Many changes concerning units and OBC in the scattering case.
Some of the problems were due to me not properly plotting the resulting quantities. But then the OBC from Kim et al.'s paper seemed to not work in the scattering framework (how to deal with the reduced dimensionality of the interacting subdomain?)

I therefore changed the OBC to use Sancho-Rubio. This is now mostly complete up until the  polarization (give or take a constant and a sign).

I will next continue with the photon solver and the self-energy.